+++ primefaces/src/main/java/org/primefaces/component/datagrid/DataGrid.java
@@ -40,6 +40,7 @@ import org.primefaces.event.data.PageEvent;
 import org.primefaces.model.LazyDataModel;
 import org.primefaces.util.ComponentUtils;
 import org.primefaces.util.Constants;
+import org.primefaces.util.LangUtils;
 import org.primefaces.util.MapBuilder;
 
 @ResourceDependency(library = "primefaces", name = "components.css")
@@ -117,8 +118,10 @@ public class DataGrid extends DataGridBase {
                 int rows = getRowsToRender();
                 int first = Integer.parseInt(params.get(clientId + "_first"));
                 int page = rows > 0 ? (int) (first / rows) : 0;
+                String rowsPerPageParam = params.get(clientId + "_rows");
+                Integer rowsPerPage = LangUtils.isNotBlank(rowsPerPageParam) ? Integer.parseInt(rowsPerPageParam) : null;
 
-                PageEvent pageEvent = new PageEvent(this, behaviorEvent.getBehavior(), page);
+                PageEvent pageEvent = new PageEvent(this, behaviorEvent.getBehavior(), page, rowsPerPage);
                 pageEvent.setPhaseId(behaviorEvent.getPhaseId());
 
                 super.queueEvent(pageEvent);